goodix_fp_dump: improve communication with the device
Most of the details about the communication between the host and the
device have been figured out, like packets formats and checksumming
scheme.

Clean up the code and add a mechanism to communicate with the device by
sending the request type and content, and getting the response back
abstracting all the details about the packet formats.

The core function for talking with the device is now send_packet() and
an example of use is provided in send_simple_packet() and
get_msg_a8_firmware_version().