goodix_fp_dump: improve payload_memcpy and rename it to extract_payload
Make the code of payload_memcpy more robust, and make it more readable
by using buffer iterators instead of explicit offsets to move through
the source and destination memory.

Change its signature to have a packet as a parameter because this
reflects more clearly the purpose of the function: extracting the
payload from a packet.

Since the signature does no longer match that of a memcpy, change the
function name too, and take the chance to also extract the checksum in
it, so that the caller does not have to know about extra packets.

Remove the MIN() macro which is no longer needed.