goodix_fp_dump: generalize how the request data is sent
Implement a send_payload function which fragments the data into multiple
output packets if it does not fit into a single packet.

This is needed because some packets send long payloads, but the output
transfers are always 64 bytes long according to  what the Windows driver
does.

The new function is general enough to handle the case of a single packet
as well, so always use it in send_packet_full.