goodix_fp_dump: support different devices
Add a goodix_fp_device abstraction to hold some device-specific
properties like the USB device ids, the USB configuration, and the
communication endpoints.

Add  a list of (supposedly) supported devices, with the relative
properties.

Finally add support for scanning different device types and adjust the
rest of the code to handle that.